Add a common python http service library, currently a thin wrapper over Chalice
