+++ Makefile
@@ -335,7 +335,7 @@ down: ## docker-compose down - both stops and removes the containers
 	# spins up in our network, but that docker-compose doesn't know
 	# about. This must be the network that is used in Localstack's
 	# LAMBDA_DOCKER_NETWORK environment variable.
-	-docker kill $(shell docker ps --quiet --filter=network=grapl-network)
+	-docker kill $(shell docker ps --quiet --filter=network=grapl-network) &
 	docker-compose $(EVERY_COMPOSE_FILE) down --timeout=0
 	docker-compose $(EVERY_COMPOSE_FILE) --project-name $(COMPOSE_PROJECT_INTEGRATION_TESTS) down --timeout=0
 	docker-compose $(EVERY_COMPOSE_FILE) --project-name $(COMPOSE_PROJECT_E2E_TESTS) down --timeout=0
