+++ src/python/grapl-dgraph-ttl/requirements.txt
@@ -1,3 +1,4 @@
 pydgraph==20.7.0
 grapl_analyzerlib
 chalice
+grapl_http_service
