+++ src/python/grapl-http-service/BUILD
@@ -1,0 +1,11 @@
+grapl_python_distribution(
+    dependencies = [
+        "./grapl_http_service",
+    ],
+    provides = setup_py(
+        name = "grapl-http-service",
+        version = "0.1.0", # TODO: remove VERSION file? Or read from it?
+        description = "Library for building Grapl HTTP Services"
+    )
+)
+
