+++ src/python/grapl-http-service/VERSION
@@ -1,0 +1,1 @@
+0.1.0
