+++ src/python/grapl-http-service/grapl_http_service/BUILD
@@ -1,0 +1,2 @@
+python_library()
+
