+++ src/python/grapl-http-service/integration_tests/BUILD
@@ -1,0 +1,4 @@
+python_tests(
+    name="integration",
+    tags=["integration_test"],
+)
