+++ src/python/grapl-http-service/integration_tests/test_grapl_http_service.py
@@ -1,0 +1,9 @@
+import unittest
+
+import pytest
+
+@pytest.mark.integration_test
+class TestIntegrationDefault(unittest.TestCase):
+    def test_integration_grapl_http_service(self) -> None:
+        raise NotImplementedError
+
