+++ src/python/grapl-http-service/requirements.txt
@@ -1,0 +1,6 @@
+pytest
+pytest-cov
+hypothesis
+pyjwt
+chalice
+boto3_stubs
