+++ src/python/grapl-http-service/tests/BUILD
@@ -1,0 +1,4 @@
+python_tests(
+    name="unit",
+    tags=["unit_test"]
+)
