+++ src/python/grapl-http-service/tests/test_grapl_http_service.py
@@ -1,0 +1,9 @@
+import unittest
+
+import pytest
+
+@pytest.mark.unit_test
+class TestDefault(unittest.TestCase):
+    def test_grapl_http_service(self) -> None:
+        raise NotImplementedError
+
