+++ src/python/grapl-ux-router/requirements.txt
@@ -2,5 +2,5 @@ pydgraph==20.7.0
 boto3
 pyjwt==2.0.*
 grapl_analyzerlib
-chalice==1.22.4
 grapl_common
+grapl_http_service
